Use two-byte instead of byte for memory

+++ include/CHIP-8/CHIP-8.h
@@ -9,7 +9,7 @@ namespace CHIP8
 	class CHIP8
 	{
 	public:
-		uint8_t ReadMemory(const uint16_t address) const;
+		uint16_t ReadMemory(const uint16_t address) const;
 		void WriteMemory(const uint16_t address, const uint8_t data);
 
 		void SetPixel(const uint8_t row, const uint8_t col);
